+++ packages/backend/src/@types/knex-tables.d.ts
@@ -7,11 +7,13 @@ import {
     OrganizationTable,
 } from '../database/entities/organizations';
 import { UserTable, UserTableName } from '../database/entities/users';
+import { EmailTable, EmailTableName } from '../database/entities/emails';
 
 declare module 'knex/types/tables' {
     interface Tables {
         [InviteLinkTableName]: InviteLinkTable;
         [OrganizationTableName]: OrganizationTable;
         [UserTableName]: UserTable;
+        [EmailTableName]: EmailTable;
     }
 }